rename : coverLetter/edit 구조 조정
